bluetooth/gatt: fix various issues
The GATT server/client example was no longer working after
implementation of the new init/deinit mechanism in the underlying
C/C++ layers. It appears that node.js garbage collector releases
the GATT objects even when the loop is still running, which causes
the test program to fail.

To remediate to this issue, the following changes have been made:

 - Rework the GATT server API to have the
   Service/Characteristic/Descriptor objects tied to the main GATT
   server object.
 - Modify the GATT server example to make the GATT server object
   global, which prevents the garbage collector from releasing it
   wrongfully. This solution works for the example but should be
   adapted depending on application handling of the GATT server object.

Change-Id: Ief067c79182b48f9bf54dd7658116ec33713fc36
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>