lwm2m: fix wrong SSL configuration initialization
The SSL configuration structure of the LWM2M client was initialized
with 'new' but was not zeroized. This would lead to an error when
the calling application did not specify a server certificate, which
is actually supported in the case we use 'verify_cert=none'. This patch
uses malloc+memset to properly initialize this C structure.

Also the LWM2M example program has been modified to use the TCP/TLS
endpoint of ARTIK Cloud, and use a different SSL configuration: no
client certificate + no server certificate verification.

Change-Id: I22aef7d5cb405fb455d1010f2206b2644a671f33
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>