security: add JS layer and PKCS7 verification API
This patch adds a JS layer to the security module and expose
and API for doing PKCS7 signature verification. This API is
called by the 'pkcs7-verify-signature.js' command-line tool
instead of implementing the whole logic

The security test is also modify to comply with the new JS layer
instantiation.

Change-Id: I4c1accab3f72de89882363763d195a8444c9baad
Signed-off-by: Geoffroy Letourneur <g.letourneur@samsung.com>
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>