http: rework 'get_stream' function to use the proper C API
The JS 'get_stream' API used to launch the synchronous C version.
This patch modifies this behavior for the JS API to call the C
asynchronous function and actually yield back to the caller and not
block the calling thread. It also modifies the associated test
and example to work with the new behavior.

Change-Id: I1e0203a8f69b747b6a9326d5972b569674ea2c00
Signed-off-by: Geoffroy Letourneur <g.letourneur@samsung.com>
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>